Add files via upload
The project utilizes various pre-trained CNN models, such as EfficientNet and ResNet, for image classification tasks. These models, pre-trained on large datasets like ImageNet, are being fine-tuned for a specific task. A dataset is built from images stored in directories, and the model is trained on this dataset with adjustments in parameters like batch size and image size. The trained model is then saved for further use, focusing on optimizing performance in the classification task.
Load the Pretrained Model: Pretrained versions of ResNet are readily available in frameworks like TensorFlow and PyTorch.